Add investment downloads
This commit wraps existing dictionaries in a wrapper dictionary.  This
allows us to store more than one plaid transaction response in the
output file.  The requests_types field at the root states which
responses are included in the json file

We now store transactions, investment_transactions, and
investment_holdings as possible outputs for an account.

The account_type field in the settings.json file specifies which type of
transactions the download.py tool will download